bpf: switch egress gateway logic to identity_is_cluster()
Currently to determine if a given destination IP belongs to the cluster,
the egress gateway datapath logic uses is_cluster_destination(), which
relies on different sources of truth (tunnel_endpoint, security ID and EP
map).

This commit switches that logic to the new is_cluster_destination(),
which relies solely on security IDs.

Fixes: #19784

Signed-off-by: Gilberto Bertin <jibi@cilium.io>